+++ .travis-build.sh
@@ -35,8 +35,8 @@ BIN_DEVEL_DIR="${SCRIPT_DIR}/checker/bin-devel/"
 
 # For debugging
 (cd "$CHECKERFRAMEWORK" && ./gradlew --console=plain getPlumeScripts)
-"${BIN_DEVEL_DIR}/plume-scripts/ci-info" typetools
-eval "$("${BIN_DEVEL_DIR}/plume-scripts/ci-info" typetools)"
+"${BIN_DEVEL_DIR}/.plume-scripts/ci-info" typetools
+eval "$("${BIN_DEVEL_DIR}/.plume-scripts/ci-info" typetools)"
 
 source "$BIN_DEVEL_DIR/checker/bin-devel/clone-related.sh"